Tooltips
Used Bootstrap to create tooltips by adding HTML data and title attributes to an element. Used a couple of ways to optionally configure a Tooltip to use HTML inside it, and to control its placement around the element. 
Added a small jQuery script to the page to activate tooltips for that page